requirements.txt
google-genai
streamlit
streamlit-webrtc
av
pillow
python-dotenv